+++ src/pages/Qna.tsx
@@ -4,6 +4,7 @@ import { Link } from "react-router-dom";
 import { db } from "../firebase";
 import { Layout } from "@/components/layout/Layout";
 import { getAuth, onAuthStateChanged } from "firebase/auth";
+import QuestionModal from "@/components/modal/QuestionModal";
 
 type Question = {
   id: string;
@@ -17,60 +18,6 @@ type QuestionList = {
   author: string;
   authorUid: string;
 }[];
-
-const QuestionModal = ({
-  isOpen,
-  onClose,
-  onSubmit,
-}: {
-  isOpen: boolean;
-  onClose: () => void;
-  onSubmit: (title: string, content: string) => void;
-}) => {
-  const [title, setTitle] = useState("");
-  const [content, setContent] = useState("");
-
-  if (!isOpen) return null;
-
-  const handleSubmit = () => {
-    onSubmit(title, content);
-    setTitle("");
-    setContent("");
-  };
-
-  return (
-    <div className="fixed inset-0 flex items-center justify-center bg-black bg-opacity-50 ">
-      <div className="p-4 bg-white rounded ">
-        <h2 className="mb-4 text-xl">질문 추가</h2>
-        <input
-          type="text"
-          value={title}
-          onChange={(event) => setTitle(event.target.value)}
-          placeholder="제목"
-          className="w-full p-2 mb-2 border"
-        />
-        <textarea
-          value={content}
-          onChange={(event) => setContent(event.target.value)}
-          placeholder="내용"
-          className="w-full p-2 mb-2 border"
-        />
-        <button
-          onClick={handleSubmit}
-          className="p-2 mr-2 text-white bg-blue-500 rounded"
-        >
-          추가
-        </button>
-        <button
-          onClick={onClose}
-          className="p-2 text-white bg-gray-500 rounded"
-        >
-          닫기
-        </button>
-      </div>
-    </div>
-  );
-};
 
 export const Qna = () => {
   const [questions, setQuestions] = useState<Question>([]);
